switch from Blake2b to WyHash
Instead of using Blake2b as part of the Feistel round function use a
significantly faster non-cryptographic hash function WyHash. We already
know that we can use a weak function in the Feistel round function and
it'll combine and produce a cryptographically strong output.

WyHash is the fastest 64-bit hash function that passes SMHasher and also
a PRNG-version that uses the same primitives passed BigCrush [1] [2].
Using WyHash is significantly faster (one order of magnitude) and still
passes the (primitive) chi-squared randomness tests.

[1] https://github.com/rurban/smhasher

[2] https://github.com/wangyi-fudan/wyhash